:white_check_mark: #12 HTTPClientのMockを作成する関数を別のファイルに分離
